Create a json_pointer_private.h and move a few things there, fix test warnings, note array_list_insert_idx is private.
