+++ json_object_private.h
@@ -100,25 +100,6 @@ void _json_c_set_last_err(const char *err_fmt, ...);
 
 extern const char *json_hex_chars;
 
-struct json_pointer_get_result {
-	struct json_object *parent;
-	struct json_object *obj;
-	union {
-		const char *key;
-		uint32_t index;
-	} id;
-};
-
-int json_pointer_get_internal(struct json_object *obj, const char *path,
-                              struct json_pointer_get_result *res);
-
-typedef int(*json_pointer_array_set_cb)(json_object *parent, size_t idx,
-                                        json_object *value, void *priv);
-
-int json_pointer_set_with_array_cb(struct json_object **obj, const char *path,
-                                   struct json_object *value,
-                                   json_pointer_array_set_cb array_set_cb, void *priv);
-
 #ifdef __cplusplus
 }
 #endif
